# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/orders/detail/{id}": {
            "get": {
                "security": [
                    {
                        "Authorization Token": []
                    }
                ],
                "description": "Obtiene el detalle de la order segun su ID Valida usuario autorizado; existencia del detalla; detalle a mostrar asociada a usuario de session para el rol cliente",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order Detail"
                ],
                "summary": "Recupera una Orden segun su ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Detail ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.OrderItems"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "Authorization Token": []
                    }
                ],
                "description": "modifica el detalle de la orden segun su ID valida  usuario autorizado; existencia del detalle; orden asociada a usuario de session para el rol cliente; existencia  del artiulo a ser editado al detalle de la orden",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order Detail"
                ],
                "summary": "Modifica una Orden segun su ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Detail ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Create detail order",
                        "name": "orderitemdto",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.OrderItemDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.OrderItems"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Authorization Token": []
                    }
                ],
                "description": "Crea el detalle de la orden valida  usuario autorizado; existencia de la orden a agregar el detalle; orden asociada a usuario de session para el rol cliente; existencia y disponibilidad del artiulo a ser agregado al detalle de la orden",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order Detail"
                ],
                "summary": "Crea registro de las ordenes de los",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Order ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Create detail order",
                        "name": "orderitemdto",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.OrderItemDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.MessageInfo"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Authorization Token": []
                    }
                ],
                "description": "Elimina una un detalle de la orden, libera disponibilidad de articulo asociado. valida  usuario autorizado; existencia del detalle; orden asociada a usuario de session para el rol cliente; existencia  del artiulo a ser eliminado al detalle de la orden.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order Detail"
                ],
                "summary": "Elimina un detalle de la Orden segun su ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Detail ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "$ref": "#/definitions/dto.MessageInfo"
                        }
                    }
                }
            }
        },
        "/api/v1/orders/details/{id}": {
            "get": {
                "security": [
                    {
                        "Authorization Token": []
                    }
                ],
                "description": "Lista detalle de la orden valida  usuario autorizado; existencia de orden; detalle de orden a mostrar asociado a usuario de session para el rol cliente",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order Detail"
                ],
                "summary": "Recupera listado de articulos de detalle de la orden",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Order ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.OrderItems"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.MessageInfo": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
        "dto.OrderItemDTO": {
            "type": "object",
            "required": [
                "item_id",
                "quantity"
            ],
            "properties": {
                "item_id": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "integer",
                    "minimum": 1
                }
            }
        },
...
        "models.OrderItems": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "item_id": {
                    "type": "integer"
                },
                "name_item": {
                    "type": "string"
                },
                "order_id": {
                    "type": "integer"
                },
                "price": {
                    "type": "number"
                },
                "quantity": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "total": {
                    "type": "number"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
    }
}